# doc/swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: Athena
  license:
    name: MIT
host: athena.io
basePath: /api/v1
schemes:
  - https
paths:
  /books:
    get:
      tags:
        - Book
      summary: 搜索书籍
      operationId: searchBooks
      parameters:
        - $ref: '#/parameters/Authorization'
        - in: query
          type: string
          name: publisher
          required: false
          description: 出版社名称
        - in: query
          name: title
          type: array
          items:
            type: string
            uniqueItems: true
          required: false
          description: 书名,如果由提供多个，则寻找包含任一题目的书籍,多个作者应该使用title=A&title=B&title=C的形式传入
        - in: query
          name: author
          type: array
          items:
            type: string
            uniqueItems: true
          required: false
          description: 作者名称，如果由提供多个，则寻找包含所有作者的书籍,多个作者应该使用author=A&author=B&author=C的形式传入
        - in: query
          name: limit
          type: integer
          minimum: 1
          description: 每页结果数量最大值
          required: false
          default: 20
        - in: query
          type: integer
          name: startPage
          minimum: 0
          description: 起始页号（从1开始）
          required: false
          default: 1
      responses:
        '200':
          description: 查找到的所有书籍
          headers:
            X-Total-Count:
              type: integer
              description: 结果总数
            Link:
              type: array
              items:
                type: string
                pattern: '<(http|https)://(.*)>;ref="(first|last|previous|next)"'
              description: 分页相关链接
          schema:
            $ref: '#/definitions/BookSearchResult'
        '403':
          $ref: '#/responses/Forbidden'
...
definitions:
  Error:
    type: object
    properties:
      code:
        type: string
      msg:
        type: string
    required:
      - code
      - msg
...
  Book:
    type: object
    description: 图书对象
    required:
      - isbn
      - title
      - language
      - publisher
      - publish_date
      - category_id
    properties:
      isbn:
        type: integer
        format: int64
        description: ISBN号
      title:
        type: string
        description: 书籍标题
      subtitle:
        type: string
        description: 副标题
      language:
        type: string
        description: 书籍语言
      publisher:
        $ref: '#/definitions/Publisher'
      publish_date:
        type: string
        description: 出版日期
      category_id:
        type: string
        description: 中图分类号
      version:
        type: integer
        description: 版本号
      cover_url:
        type: string
        description: 封面url
      preface:
        type: string
        description: 前言
      directory:
        type: object
        description: 目录(JSON格式)
      introduction:
        type: string
        description: 简介
      price:
        type: number
        format: float32
        description: 定价
  Publisher:
    type: object
    description: 出版社
    required:
      - id
      - name
      - location
    properties:
      id:
        type: string
        description: 出版社编号
      name:
        type: string
        description: 出版社名称
      location:
        type: string
        description: 出版社所在城市
  BookSearchResult:
    type: object
    description: 图书搜索结果
    properties:
      count:
        type: integer
        minimum: 1
        description: 查询结果总数
      limit:
        type: integer
        minimum: 1
        default: 20
        description: 每页查询结果限制
      offset:
        type: integer
        default: 0
        description: 已经返回了多少结果，即本次返回的第一个结果数-1
      results:
        type: array
        items:
          $ref: '#/definitions/Book'
        description: 结果
...
responses:
...
  Forbidden:
    description: 未授权
    schema:
      $ref: '#/definitions/Error'
...
parameters:
  Authorization:
    in: header
    name: X-AUTHENTICATION
    type: string
    required: true
    description: JWT鉴权码